requirements.txt
instagrapi
moviepy
pillow
PyQt5
activesoup
requests
playwright==1.26
yt_dlp
pyrogram
vk_api
scrapetube
loguru
urllib3==1.26.15
kthread
tgcrypto
python-dateutil
pyinstaller